Refactoring project (#26)
* move TransactionsController#create into a UseCase class

* adapt TransactionController methods to the use_case pattern

* split specs into smaller files

* start writing readme

* fix specs and small bugs
