+++ README.md
@@ -1,24 +1,68 @@
-# README
+# CO Budget API
+This is the API of a simple budgeting app that allows users to track their expenses and income.
 
-This README would normally document whatever steps are necessary to get the
-application up and running.
+## Main Features
+- Users can create an account and log in.
+- Users can create, read, update, and delete their budgets
+- Users can create, read, update, and delete their budget's transactions (expenses and income)
+  - Transactions can be once, installments, or recurring
 
-Things you may want to cover:
+## TODO
+- [ ] Setup Docker for the project
+- [ ] Allow users to invite other users to their budget
 
-* Ruby version
+Stack:
+- NextJS (Frontend)
+- Ruby On Rails
+- PostgreSQL
+- RSpec
+- Heroku
+- Rubocop
+- SimpleCov
 
-* System dependencies
+## Setup
+##### Prerequisites
 
-* Configuration
+The setups steps expect following tools installed on the system.
 
-* Database creation
+- Github
+- Ruby [2.1.3]
+- Rails [7.0.4]
 
-* Database initialization
+##### 1. Check out the repository
 
-* How to run the test suite
+```bash
+git clone git@github.com:LucasHFS/co-budget-api.git
+```
 
-* Services (job queues, cache servers, search engines, etc.)
+##### 2. Setup database.yml file
 
-* Deployment instructions
+Edit the database configuration as required.
 
-* ...
+##### 3. Create and setup the database
+
+Run the following commands to create and setup the database.
+
+```ruby
+bundle exec rake db:create
+bundle exec rake db:setup
+```
+
+##### 4. Start the Rails server
+
+You can start the rails server using the command given below.
+
+```ruby
+bundle exec rails s
+```
+
+## App Pages
+
+### Home
+![list](imgs/home.png)
+![new transaction](imgs/new_transaction.png)
+### Mobile
+![List (mobile)](imgs/mobile_home.png)
+![Create (mobile)](imgs/mobile_form.png)
+## Small Demo
+![small_demo](imgs/small-demo.gif)
