+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_09_23_153834) do
+ActiveRecord::Schema[7.0].define(version: 2023_11_18_145713) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -47,7 +47,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_09_23_153834) do
     t.string "name", null: false
     t.integer "price_in_cents", default: 0, null: false
     t.date "due_at", null: false
-    t.integer "status", default: 1, null: false
+    t.integer "status", default: 2, null: false
     t.integer "kind", default: 1, null: false
     t.integer "transaction_type", default: 1, null: false
     t.integer "installment_number", default: 1
@@ -58,6 +58,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_09_23_153834) do
     t.index ["collection_id"], name: "index_transactions_on_collection_id"
     t.index ["kind"], name: "index_transactions_on_kind"
     t.index ["status"], name: "index_transactions_on_status"
+    t.index ["transaction_type"], name: "index_transactions_on_transaction_type"
   end
 
   create_table "users", force: :cascade do |t|
