--- spec/factories/collections.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-FactoryBot.define do
-  factory :collection do
-    kin { 1 }
-    installment_number { 1 }
-  end
-end
