--- spec/requests/budgets_spec.rb
@@ -1,299 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-RSpec.describe 'Budgets' do
-  describe 'GET /index' do
-    subject(:request) { get '/api/budgets', params: {}, headers: }
-
-    let(:json_body) do
-      JSON.parse(response.body, symbolize_names: true)
-    end
-
-    let(:user) { create(:user) }
-
-    before do
-      create_list(:budget, 5, users: [user])
-      create_list(:budget, 3)
-    end
-
-    context 'when not logged in' do
-      let(:headers) do
-        { 'Accept' => 'application/json' }
-      end
-
-      it 'returns unauthorized' do
-        request
-
-        expect(response).to have_http_status(:unauthorized)
-      end
-    end
-
-    context 'when logged in' do
-      let(:token) { user.generate_jwt }
-      let(:headers) { { 'Authorization' => "Bearer #{token}" } }
-
-      it 'is successful' do
-        request
-        expect(response).to have_http_status(:success)
-      end
-
-      it "lists all user's budgets" do
-        request
-        expect(json_body[:budgets].length).to eq(5)
-      end
-    end
-  end
-
-  describe 'POST /create' do
-    subject(:request) { post '/api/budgets', params:, headers: }
-
-    let(:json_body) do
-      JSON.parse(response.body, symbolize_names: true)
-    end
-
-    let(:user) { create(:user) }
-
-    let(:params) do
-      {
-        budget: {
-          name: 'budget a'
-        }
-      }
-    end
-
-    context 'when not logged in' do
-      let(:headers) do
-        { 'Accept' => 'application/json' }
-      end
-
-      it 'returns unauthorized' do
-        request
-
-        expect(response).to have_http_status(:unauthorized)
-      end
-    end
-
-    context 'when logged in' do
-      let(:token) { user.generate_jwt }
-      let(:headers) { { 'Authorization' => "Bearer #{token}" } }
-
-      context 'when successful' do
-        it 'is returns :created status' do
-          request
-          expect(response).to have_http_status(:created)
-        end
-
-        it 'creates the budget on the db' do
-          request
-
-          expect(JSON.parse(response.body).with_indifferent_access).to match(
-            {
-              budget: {
-                id: Integer,
-                name: 'budget a'
-              }
-            }.with_indifferent_access
-          )
-        end
-      end
-
-      context 'when validation fails' do
-        let(:params) do
-          {
-            budget: {
-              name: nil
-            }
-          }
-        end
-
-        it 'is returns :unprocessable_entity status' do
-          request
-          expect(response).to have_http_status(:unprocessable_entity)
-        end
-
-        it 'returns the error message' do
-          request
-
-          expect(JSON.parse(response.body).with_indifferent_access).to match(
-            {
-              error: {
-                message: 'Erro de validação',
-                details: ['Name não pode ficar em branco']
-              }
-            }.with_indifferent_access
-          )
-        end
-      end
-    end
-  end
-
-  describe 'PUT /update' do
-    subject(:request) { put "/api/budgets/#{budget.id}", params:, headers: }
-
-    let(:json_body) do
-      JSON.parse(response.body, symbolize_names: true)
-    end
-
-    let(:user) { create(:user) }
-    let(:budget) { create(:budget, users:[user]) }
-
-    let(:params) do
-      {
-        budget: {
-          name: 'budget b'
-        }
-      }
-    end
-
-    context 'when not logged in' do
-      let(:headers) do
-        { 'Accept' => 'application/json' }
-      end
-
-      it 'returns unauthorized' do
-        request
-
-        expect(response).to have_http_status(:unauthorized)
-      end
-    end
-
-    context 'when logged in' do
-      let(:token) { user.generate_jwt }
-      let(:headers) { { 'Authorization' => "Bearer #{token}" } }
-
-      context 'when successful' do
-        it 'is returns :ok status' do
-          request
-          expect(response).to have_http_status(:ok)
-        end
-
-        it 'creates the budget on the db' do
-          request
-
-          created_budget = Budget.last.attributes.transform_keys(&:to_sym)
-
-          expect(created_budget).to include(
-            {
-              name: 'budget b'
-            }
-          )
-        end
-      end
-
-      context 'when validation fails' do
-        let(:params) do
-          {
-            budget: {
-              name: nil
-            }
-          }
-        end
-
-        it 'is returns :unprocessable_entity status' do
-          request
-          expect(response).to have_http_status(:unprocessable_entity)
-        end
-
-        it 'returns the error message' do
-          request
-
-          expect(JSON.parse(response.body).with_indifferent_access).to match(
-            {
-              error: {
-                message: 'Erro de validação',
-                details: ['Name não pode ficar em branco']
-              }
-            }.with_indifferent_access
-          )
-        end
-      end
-    end
-  end
-
-  describe 'DELETE /delete' do
-    subject(:request) { delete "/api/budgets/#{budget_id}", params: {}, headers: }
-
-    let(:json_body) do
-      JSON.parse(response.body, symbolize_names: true)
-    end
-
-    let(:user) { create(:user) }
-    let!(:budget) { create(:budget, name: 'budget to delete', users:[user]) }
-    let(:budget_id) { budget.id }
-
-    context 'when not logged in' do
-      let(:headers) do
-        { 'Accept' => 'application/json' }
-      end
-
-      it 'returns unauthorized' do
-        request
-
-        expect(response).to have_http_status(:unauthorized)
-      end
-    end
-
-    context 'when logged in' do
-      let(:token) { user.generate_jwt }
-      let(:headers) { { 'Authorization' => "Bearer #{token}" } }
-
-      context 'when successful' do
-        it 'is returns :no_content status' do
-          request
-          expect(response).to have_http_status(:no_content)
-        end
-
-        it 'creates the budget on the db' do
-          request
-
-          expect(Budget.find_by(name: 'budget to delete')).to be_nil
-        end
-      end
-
-      context 'when validation fails' do
-        let!(:transaction) { create(:transaction, budget:) }
-
-        it 'is returns :unprocessable_entity status' do
-          request
-          expect(response).to have_http_status(:unprocessable_entity)
-        end
-
-        it 'returns the error message' do
-          request
-
-          expect(JSON.parse(response.body).with_indifferent_access).to match(
-            {
-              error: {
-                message: 'Erro de exclusão',
-                type: 'BadRequest'
-              }
-            }.with_indifferent_access
-          )
-        end
-      end
-
-      context 'when not found' do
-        subject(:request) { delete '/api/budgets/non-existing', params: {}, headers: }
-
-        it 'is returns :unprocessable_entity status' do
-          request
-          expect(response).to have_http_status(:not_found)
-        end
-
-        it 'returns the error message' do
-          request
-
-          expect(JSON.parse(response.body).with_indifferent_access).to match(
-            {
-              error: {
-                message: 'Recurso não encontrado',
-                type: 'NotFoundError'
-              }
-            }.with_indifferent_access
-          )
-        end
-      end
-    end
-  end
-end
